fix: resolve threads final review gates
